--- crates/pessimistic-proof-core/src/utils/mod.rs
@@ -1,2 +1,0 @@
-pub mod empty_hash;
-pub mod smt;
